multiple bugfixes for region and baseline (only in hash for now)

+++ btb/geometric_xbtbx/geometric_xbtbx.cc
@@ -198,7 +198,7 @@ struct FilterBTBEntry {
   uint64_t target;
   branch_info type = branch_info::ALWAYS_TAKEN;
   std::tuple<uint16_t, uint16_t, uint64_t> region_idx_tag = {0, 0, 0};
-  uint8_t target_size = 64; // TODO: Only update for which we have sizes
+  uint16_t target_size = 64; // TODO: Only update for which we have sizes
   uint64_t offset_mask = -1;
 
   // TODO: shift indexes and tags into place
@@ -231,7 +231,7 @@ struct BTBEntry {
   branch_info type = branch_info::ALWAYS_TAKEN;
   // Set / precise / magic pointers
   std::tuple<uint16_t, uint16_t, uint64_t> region_idx_tag = {0, 0, 0};
-  uint8_t target_size = 64; // TODO: Only update for which we have sizes
+  uint16_t target_size = 64; // TODO: Only update for which we have sizes
   uint64_t offset_mask = -1;
   uint8_t precise_branch_type;
   bool useless = false;
